FPU context saving/restoring enabled. Do not use lazy fpu saving at the moment.
git-svn-id: https://svn.aros.org/svn/aros/trunk@31166 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/ppc-chrp/efika/include/asm/mpc5200b.h
@@ -21,7 +21,7 @@ typedef struct regs {
 
 typedef struct fpuregs {
     double      fpr[32];
-    uint32_t    fpscr;
+    uint64_t    fpscr;
 } fpuregs_t;
 
 typedef struct context {
